# api_contract.yaml
openapi: 3.1.0
info:
  title: Derived API Contract
  version: 1.0.0
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        currency:
          type: string
        imageUrl:
          type: string
        stock:
          type: integer
        categoryId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AuthToken:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
...
paths:
  /api/auth/login:
    post:
      operationId: post_api_auth_login
      responses:
        '201':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...
  /api/auth/refresh:
    post:
      operationId: post_api_auth_refresh
      responses:
        '201':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...
  /api/categories:
    get:
      operationId: get_api_categories
      responses:
        '200':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      operationId: post_api_categories
      responses:
        '201':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /api/categories/:id:
    delete:
      operationId: delete_api_categories_id
      responses:
        '204':
          description: Derived from SPEC.md
    put:
      operationId: put_api_categories_id
      responses:
        '200':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...
  /api/products:
    get:
      operationId: get_api_products
      responses:
        '200':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      operationId: post_api_products
      responses:
        '201':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
  /api/products/:id:
    delete:
      operationId: delete_api_products_id
      responses:
        '204':
          description: Derived from SPEC.md
    get:
      operationId: get_api_products_id
      responses:
        '200':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    put:
      operationId: put_api_products_id
      responses:
        '200':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object